Rafael

+++ src/app/features/customers/components/list-customers/list-customers.component.ts
@@ -7,6 +7,8 @@ import { heroPencilSquare, heroTrash } from '@ng-icons/heroicons/outline';
 import { ModalService } from '@app/shared/services/modal.service';
 import { UpdateCustomerComponent } from '../update-customer/update-customer.component';
 import { BreakpointObserver } from '@angular/cdk/layout';
+import { ConfirmDialogComponent } from '@app/shared/components/confirm-dialog/confirm-dialog.component';
+import { AlertService } from '@app/shared/services/alert.service';
 
 @Component({
   selector: 'app-list-customers',
@@ -17,10 +19,11 @@ import { BreakpointObserver } from '@angular/cdk/layout';
   styleUrl: './list-customers.component.scss',
 })
 export class ListCustomersComponent {
-  private modalService = inject(ModalService)
+  private _modalService = inject(ModalService);
   private _customerService = inject(CustomersService);
-  customers: CustomerResponse[] = [];
   private breakpointObserver = inject(BreakpointObserver);
+  private _alertService = inject(AlertService);
+  customers: CustomerResponse[] = [];
 
   isMobile = signal(false);
 
@@ -34,26 +37,36 @@ export class ListCustomersComponent {
       this.customers = customers;
       console.log(customers);
     });
-
+    console.table(this.customers);
     this.updateListCustomers();
+    console.table(this.customers);
   }
 
   updateListCustomers() {
-    this._customerService.findAll().subscribe();
+    this._customerService.searchCustomers('', '1', '10').subscribe();
   }
 
   deleteCustomer(id: number) {
-    /*Preguntar si quiere eliminarlo uwu */
-    this._customerService.deleteCustomerById(id).subscribe({
-      next: () => {
-        /**Alerta success */
-      }
+    console.log(id);
+    this._modalService
+    .openWithResult(ConfirmDialogComponent, {}, {
+      message: '¿Estás seguro de que querés eliminar este cliente?'
     })
+    .subscribe((confirmed: boolean) => {
+      console.log(confirmed);
+      if (confirmed) {
+        this._customerService.deleteCustomerById(id).subscribe({
+          next: () => {
+            this._alertService.showSuccess('Cliente eliminado');
+          }
+        });
+      }
+    });
   }
 
   updateCustomer(customer: CustomerResponse) {
     this._customerService.customerForUpdate.next(customer);
-    this.modalService.open(UpdateCustomerComponent);
+    this._modalService.open(UpdateCustomerComponent);
   }
 
 }
